requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
rest_framework_simplejwt